Use 'uname' to check the OS we are running on
In order to support non-Linux operating systems, we have to know what
operating system we are using. As there is no function in GLib and gjs
can be use to get the name of the running OS, we rely on external
'uname' command to do this. The result is stored in 'kernel' variable.