Make smDialog message changeable and don't require NetworkManager on FreeBSD
Now we have different types of errors: operating system not supported,
code not compiled. To make showing error messages easier, we make
message field in smDialog changeable, so other types of errors can
reuse the code.

NetworkManager is only available on Linux, so we cannot require it on
other systems. This means Net element is currently broken on FreeBSD
and you have to comment it out to test this extension on FreeBSD.