Port Freq element to FreeBSD
CPU frequency information is available on FreeBSD through sysctl
interface. As it is too hard to make a gir to call sysctl directly
in gjs, I decided to make a wrapper for it.

This commit also adds a Makefile to ease compilation of wrapper code
because users are required to compile it if they don't install this
extension from the package manager. We are able to load typelib from
extension directory now, and error messages are showed if the loading
of system-dependent wrapper fails.