Call glibtop_init if it is available
Calling glibtop_init to make sure glibtop_global_server is correctly
initialized. Without calling this function, the global server instance
may use wrong initialization flags and calling glibtop_get_swap aborts
gnome-shell on FreeBSD.

Unfortunately, introspection binding in libgtop was not done correctly,
so glibtop_init was not available through introspection because
introspectable="0" is set. This problem has been fixed in this commit:
https://git.gnome.org/browse/libgtop/commit/?id=598764c

The fix is available in libgtop 2.31.4 and later versions. To keep this
extension compatible with older version of libgtop, we only calls
glibtop_init when it is available.

The fix is also backported to FreeBSD ports:
https://svnweb.freebsd.org/changeset/ports/394176

Therefore, all FreeBSD users who want to use this extension must upgrade
to libgtop-2.30.0_5 or later versions before installing it.