Port Net element to FreeBSD
The list of network interfaces is available on FreeBSD through
getifaddrs(3), but making binding for this function and all required
struct and constant definitions will be a hard work. Therefore, I made
a wrapper to get the list and ignore loopback and bridge interfaces.

NetworkManager is not available on non-Linux systems, so all code using
NetworkManager functions is moved to Linux-specific functions. To get
notified when states of interfaces are changed, we connect to FreeBSD
devd socket to get devctl(4) events.

This commit also makes the system-dependent wrapper code depend on GLib
because we need some easy-to-use data structures.