Improve cpu-frequency graph range.
The graph is now scaled to the range min-frequency to max-frequency,
i.e. when the scaling is at min-frequency the graph now shows 0. The
displayed values (menu etc) still show actual frequence, though.

This improves the usefulness of the graph because A) the range from min
to max frequency is often such that the graph only varies in the top 30%
or 40%, and B) the graph now looks more like the cpu-usage graph, i.e.
when idling both are near 0 and when busy both are near 100%.