prefs: Remove clutter dependency on GNOME 40
There is a strict distinction between GNOME Shell using the Clutter toolkit
and apps (including extension preferences) using GTK.

And while it is technically possible for apps to use Clutter, the library
is unmaintained and its use is discouraged. Additionally, GNOME Shell
actually uses its own internal fork of Clutter so if we depend on Clutter,
we need to install a second copy of it just for the preferences.

Fortunately, GDK 4 supports parsing hex colours with alpha values
(https://gitlab.gnome.org/GNOME/gtk/merge_requests/2356),
so we can just use it directly without depending on Clutter.

We will only use the Compat module when GTK 4 is not available,
such as on GNOME < 40.