Optionally hide on lockscreen while continuing to produce graphs in background
Hidden by default.

This makes use of the session modes feature introduced in GNOME Shell 42.

https://gjs.guide/extensions/topics/session-modes.html

Strangely, disconnecting the session mode updated callback, like in the example, resulted in the applet still being shown the first time the screen is locked. It seems to work fine with it left out, so I've just done that.