actions: uploader: Only submit new builds when we get new tags
Currently we submit a new build anytime anything is pushed to the master
branch. Unfortunately this means we spam a bunch of builds to
extensions.gnome.org if a multiple commits are pushed (if a multi-commit PR
is merged, for example).

Make build submission a bit more explicit by only running the extension
uploader action on new tags.