extension: Don't recommend Alt-F2 r
There is currently a fantastically annoying bug in Gnome Shell [1] that
results in a crash if you do an `r' with *any* windows open, so people are
crashing their systems by following the instructions in our
description. Drop our gs restart recommendation for now (and hopefully
forever since we should really just fix the error instead of recommending a
gs restart #50).

Fixes #49

[1] https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/7050