extension: Drop gtop and nm dependency checks and pop-up dialog
Our system dependency pop-up dialog (smDialog) is currently using some
private APIs which are being removed in Gnome Shell 46 [1]. Rather than
fixing this let's just tie it up and throw it overboard. This is a
reasonable course of action because:

  (1) It's not clear which replacement APIs we should be moving to. We
      could figure this out with a bit of research and testing, but it
      may not be worth it because:

  (2) The smDialog dependency checks have actually been hard-coded out
      since the port to Gnome Shell 45 [2] (since the old dependency
      check method relied on a try/catch of the library imports, which
      isn't supported in es6).

So even if we figure out a new API for copying props to the child, we'd
still never be using any of that new code anyway since we're hard-coding an
assumption of the dependencies being present.

Ultimately we *should* fix this for the best possible user experience, but
in the interest of keeping things moving with Gnome Shell updates we'll
defer that work to the back burner (#69).

[1] https://github.com/mgalgs/gnome-shell-system-monitor-applet/pull/64#issuecomment-1986346950
[2] (f838b9786: "extension: Update for Gnome Shell 45")